requirements.txt
# AC_OTFlex_Workflow Requirements
# Python 3.8+

# Core dependencies
requests>=2.28.0          # HTTP client for Opentrons API
pyserial>=3.5             # Arduino serial communication
pandas>=1.4.0             # Data handling and CSV processing
numpy>=1.23.0             # Numerical computing

# xArm robot control (optional but recommended)
xarm-python-sdk>=1.13.0   # UFactory xArm SDK

# Development and testing (optional)
pytest>=7.0               # Testing framework
pytest-asyncio>=0.20.0    # Async test support
black>=22.0               # Code formatter
flake8>=4.0               # Code linter
mypy>=0.990               # Type checker
